The start of a bit of an overhaul
This started off as a concept of grouped temp/time readings as well as grouping steps into recipes for the brew day. Now it's a complete rethink of how a brew day actually works and what else the app can do to provide extra info to  the user.

This may take some time...